document-portal: Use libfuse3 API
Use the "new" fuse API that is becoming the default given that libfuse2
is not maintained anymore.

No logic changes are done here, and all is preserved as it used to be,
we require capabilities through the init vfunc though, as some
capabilities can't be requested anymore via fuse_args.

Fixes: #194