document-portal: Use libfuse3 API
Use the "new" fuse API that is becoming the default given that libfuse2
is not maintained anymore.

No major logic changes are done here, and all is preserved as it used to
be, but we ensure that the fuse session is created and destroyed in the
same thread.
Also, we require capabilities through the init vfunc, as some features
can't be requested anymore via fuse_args.

Fixes: #194