tests: Ensure that dbus-daemon won't block the tests execution
Some tests are failing when G_MESSAGES_DEBUG=all is used because the
dbus-daemon output may cause stdout not full and then not writable
anymore. Causing a deadlock during tests.

To avoid this, in older versions of GLib explicitly make dbus-daemon to
print to another fd (mapped to stdout) and redirect the normal output
to stderr, so that glib won't have to handle that.